Add getFreeNodeCount to include/task_scheduler.h and remove extern references
Moved getFreeNodeCount declaration from extern references in src/2DD40.c
and src/3E160.c to the proper header file include/task_scheduler.h.
Changed return type from s16 to s32 to match the extern declarations
and ensure matching binary output.

+++ src/task_scheduler.c
@@ -438,7 +438,7 @@ void *scheduleTask(void *callback, u8 nodeType, u8 identifierFlag, u8 priority) 
     return NULL;
 }
 
-s16 getFreeNodeCount(s32 arg0) {
+s32 getFreeNodeCount(s32 arg0) {
     return gActiveScheduler->counters[arg0];
 }
